# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is for BRAC-Allumni-Association - A REST API for BRAC-Allumni-Association",
    "version": "1.0.6",
    "title": "BRAC-Allumni-Association"
  },
  "host": "localhost:5050",
  "basePath": "/api/v1",
...
  "schemes": ["http", "https"],
  "paths": {
    "/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Create user",
        "description": "This can only be done by the logged in user who has admin power.",
        "operationId": "createUser",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "firstName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "confirmPassword",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "contactNumber",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "required": true,
            "type": "string",
            "enum": ["male", "female", "other"]
          },
          {
            "name": "photo",
            "in": "formData",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": ["Users"],
        "summary": "get all users",
        "description": "Anyone can access this route to signup",
        "operationId": "getAllUsers",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "number"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/users/update-me": {
      "patch": {
        "tags": ["Authentication"],
        "summary": "update login user",
        "description": "Anyone can access this route to signup",
        "operationId": "updateMe",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "firstName",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "lastName",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "contactNumber",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "required": false,
            "type": "string",
            "enum": ["male", "female", "other"]
          },
          {
            "name": "photo",
            "in": "formData",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/users/delete-me": {
      "delete": {
        "tags": ["Authentication"],
        "summary": "delete login user",
        "description": "Anyone can access this route to signup",
        "operationId": "deleteMe",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/approve/{id}": {
      "patch": {
        "tags": ["Users"],
        "summary": "approve user",
        "description": "Anyone can access this route to approve user",
        "operationId": "approveUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id to approve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}